+++ backend/src/main/java/com/swipelab/analytics/api/AnalyticsController.java
@@ -2,11 +2,15 @@ package com.swipelab.analytics.api;
 
 import com.swipelab.analytics.application.AnalyticsService;
 import com.swipelab.analytics.dto.*;
+import com.swipelab.dto.response.UserPerformanceResponse;
 import lombok.RequiredArgsConstructor;
 import org.springframework.http.ResponseEntity;
+import org.springframework.security.access.prepost.PreAuthorize;
 import org.springframework.security.core.annotation.AuthenticationPrincipal;
 import org.springframework.security.core.userdetails.UserDetails;
 import org.springframework.web.bind.annotation.*;
+
+import java.util.List;
 
 import java.time.LocalDateTime;
 import java.util.Map;
@@ -22,6 +26,7 @@ public class AnalyticsController {
     // /api/v1/statistics/progress?)
     // User request: GET /api/v1/classifications/progress
     @GetMapping("/api/v1/classifications/progress")
+    @PreAuthorize("hasAnyRole('USER', 'ADMIN')")
     public ResponseEntity<UserProgressResponse> getProgress(@AuthenticationPrincipal UserDetails userDetails) {
         return ResponseEntity.ok(analyticsService.getUserProgress(userDetails.getUsername()));
     }
@@ -29,16 +34,19 @@ public class AnalyticsController {
     // 2. User Statistics (Base Path: /api/v1/statistics)
 
     @GetMapping("/api/v1/statistics/me")
+    @PreAuthorize("hasAnyRole('USER', 'ADMIN')")
     public ResponseEntity<UserStatisticsResponse> getUserStatistics(@AuthenticationPrincipal UserDetails userDetails) {
         return ResponseEntity.ok(analyticsService.getUserStatistics(userDetails.getUsername()));
     }
 
     @GetMapping("/api/v1/statistics/me/vs-experts")
+    @PreAuthorize("hasAnyRole('USER', 'ADMIN')")
     public ResponseEntity<UserVsExpertsResponse> getUserVsExperts(@AuthenticationPrincipal UserDetails userDetails) {
         return ResponseEntity.ok(analyticsService.getUserVsExperts(userDetails.getUsername()));
     }
 
     @GetMapping("/api/v1/statistics/me/vs-users")
+    @PreAuthorize("hasAnyRole('USER', 'ADMIN')")
     public ResponseEntity<UserVsExpertsResponse> getUserVsUsers(@AuthenticationPrincipal UserDetails userDetails) {
         // Re-using same response structure or method for now as per confusion
         // Or implement separate logic.
@@ -46,11 +54,13 @@ public class AnalyticsController {
     }
 
     @GetMapping("/api/v1/statistics/me/breakdown")
+    @PreAuthorize("hasAnyRole('USER', 'ADMIN')")
     public ResponseEntity<PerformanceBreakdownResponse> getBreakdown(@AuthenticationPrincipal UserDetails userDetails) {
         return ResponseEntity.ok(analyticsService.getPerformanceBreakdown(userDetails.getUsername()));
     }
 
     @GetMapping("/api/v1/statistics/me/timeseries")
+    @PreAuthorize("hasAnyRole('USER', 'ADMIN')")
     public ResponseEntity<TimeSeriesResponse> getTimeSeries(
             @AuthenticationPrincipal UserDetails userDetails,
             @RequestParam(defaultValue = "accuracy") String metric,
@@ -58,8 +68,11 @@ public class AnalyticsController {
         return ResponseEntity.ok(analyticsService.getTimeSeries(userDetails.getUsername(), metric, period));
     }
 
+    // ADMIN ANALYTICS SECTION
+
     // 3. Task Analytics (Base Path: /dashboard)
-    @GetMapping("/dashboard/tasks/{taskId}/analytics")
+    @GetMapping("api/v1/analytics/tasks/{taskId}")
+    @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<TaskAnalyticsResponse> getTaskAnalytics(
             @PathVariable Long taskId,
             @RequestParam(required = false) Boolean includePerSpecies,
@@ -67,7 +80,8 @@ public class AnalyticsController {
         return ResponseEntity.ok(analyticsService.getTaskAnalytics(taskId));
     }
 
-    @PostMapping("/dashboard/exports")
+    @PostMapping("api/v1/analytics/exports")
+    @PreAuthorize("hasRole('ADMIN')")
     public ResponseEntity<Map<String, Object>> createExport(@RequestBody Map<String, Object> request) {
         // Placeholder for export
         return ResponseEntity.accepted().body(Map.of(
@@ -75,5 +89,17 @@ public class AnalyticsController {
                 "status", "QUEUED",
                 "createdAt", LocalDateTime.now(),
                 "estimatedCompletion", LocalDateTime.now().plusMinutes(10)));
+    }
+
+    @GetMapping("api/v1/analytics/users")
+    @PreAuthorize("hasRole('ADMIN')")
+    public List<UserPerformanceResponse> getUserPerformance(@RequestParam(required = false) Long taskId) {
+        return analyticsService.getUserPerformanceMetrics(taskId);
+    }
+
+    @GetMapping("api/v1/analytics/top-performers")
+    @PreAuthorize("hasRole('ADMIN')")
+    public List<UserPerformanceResponse> getTopPerformers(@RequestParam(defaultValue = "10") int limit) {
+        return analyticsService.getTopPerformers(limit);
     }
 }
